return float instead of int

+++ app/Models/Order.php
@@ -239,7 +239,7 @@ class Order extends Model
         return ($this->grand_total_to_be_paid) * 100;
     }
 
-    public function getGrandTotalToBePaidAttribute(): int
+    public function getGrandTotalToBePaidAttribute(): float
     {
         return $this->grand_total - $this->amount_paid_from_wallet;
     }
